+++ LoveBird/Sources/Components/TouchableView.swift
@@ -1,5 +1,5 @@
 //
-//  TouchableStack.swift
+//  TouchableView.swift
 //  LoveBird
 //
 //  Created by 황득연 on 2023/06/04.
@@ -7,7 +7,7 @@
 
 import SwiftUI
 
-struct TouchableStack<Content: View>: View {
+struct TouchableView<Content: View>: View {
   let content: Content
   
   init(@ViewBuilder content: () -> Content) {
